Move "DB" handling to another class
